refactor(prompts): Reorganize phoenix js client exports
