+++ js/packages/phoenix-client/package.json
@@ -9,6 +9,22 @@
     ".": {
       "import": "./dist/esm/index.js",
       "require": "./dist/src/index.js"
+    },
+    "./prompts": {
+      "import": "./dist/esm/prompts/index.js",
+      "require": "./dist/src/prompts/index.js"
+    },
+    "./experimental": {
+      "import": "./dist/esm/experimental/index.js",
+      "require": "./dist/src/experimental/index.js"
+    },
+    "./utils/*": {
+      "import": "./dist/esm/utils/*.js",
+      "require": "./dist/src/utils/*.js"
+    },
+    "./types/*": {
+      "import": "./dist/esm/types/*.d.ts",
+      "require": "./dist/src/types/*.d.ts"
     }
   },
   "files": [
@@ -30,6 +46,7 @@
   "license": "ELv2",
   "devDependencies": {
     "@types/node": "^20.14.11",
+    "openai": "^4.77.0",
     "openapi-typescript": "^7.4.1",
     "tsx": "^4.19.1"
   },
@@ -41,5 +58,8 @@
   "packageManager": "pnpm@9.15.0",
   "engines": {
     "node": ">=18"
+  },
+  "peerDependencies": {
+    "openai": "^4.77.0"
   }
 }
