+++ js/packages/phoenix-client/src/experimental/experiments/runExperiment.ts
@@ -1,5 +1,5 @@
-import { Dataset, Example } from "../types/datasets";
-import { createClient, type PhoenixClient } from "../client";
+import { Dataset, Example } from "../../types/datasets";
+import { createClient, type PhoenixClient } from "../../client";
 import type {
   Evaluator,
   Experiment,
@@ -8,16 +8,12 @@ import type {
   ExperimentRun,
   ExperimentTask,
   RanExperiment,
-} from "../types/experiments";
-import { promisifyResult } from "../utils/promisifyResult";
+} from "../../types/experiments";
+import { promisifyResult } from "../../utils/promisifyResult";
 import invariant from "tiny-invariant";
-import { pluralize } from "../utils/pluralize";
-
-export type Logger = {
-  info: (message: string) => void;
-  error: (message: string) => void;
-  log: (message: string) => void;
-};
+import { pluralize } from "../../utils/pluralize";
+import { getDatasetLike } from "../../utils/getDatasetLike";
+import { type Logger } from "../../types/logger";
 
 export type RunExperimentParams = {
   /**
@@ -35,6 +31,7 @@ export type RunExperimentParams = {
    */
   projectName?: string;
   logger?: Logger;
+  readonly?: boolean;
 };
 
 /**
@@ -49,9 +46,10 @@ export async function runExperiment({
   repetitions = 1,
   projectName = "default",
   logger = console,
+  readonly = false,
 }: RunExperimentParams): Promise<RanExperiment> {
   const client = _client ?? createClient();
-  const dataset = await getDataset({ dataset: _dataset, client });
+  const dataset = await getDatasetLike({ dataset: _dataset, client });
   invariant(dataset, `Dataset not found`);
   invariant(dataset.examples.length > 0, `Dataset has no examples`);
   const experimentName =
@@ -68,6 +66,12 @@ export async function runExperiment({
     repetitions,
     projectName,
   };
+
+  if (readonly) {
+    logger.info(
+      `🔧 Running experiment in readonly mode. Results will not be recorded.`
+    );
+  }
 
   logger.info(
     `🧪 Starting experiment "${experimentName}" on dataset "${dataset.id}" with task "${task.name}" and ${evaluators?.length ?? 0} ${pluralize(
@@ -160,7 +164,10 @@ function runTask({
     try {
       const taskOutput = await promisifyResult(task(example));
       // TODO: why doesn't run output type match task output type?
-      thisRun.output = JSON.stringify(taskOutput);
+      thisRun.output =
+        typeof taskOutput === "string"
+          ? taskOutput
+          : JSON.stringify(taskOutput);
     } catch (error) {
       thisRun.error = error instanceof Error ? error.message : "Unknown error";
     }
@@ -189,7 +196,10 @@ export async function evaluateExperiment({
   logger: Logger;
 }): Promise<RanExperiment> {
   const client = _client ?? createClient();
-  const dataset = await getDataset({ dataset: experiment.datasetId, client });
+  const dataset = await getDatasetLike({
+    dataset: experiment.datasetId,
+    client,
+  });
   invariant(dataset, `Dataset "${experiment.datasetId}" not found`);
   invariant(
     dataset.examples.length > 0,
@@ -290,48 +300,6 @@ async function runEvaluator({
   };
 
   return evaluate();
-}
-
-/**
- * Return a dataset object from the input.
- *
- * If the input is a string, assume it is a dataset id and fetch the dataset from the client.
- * If the input is an array of examples, create a new dataset from the examples then return it.
- * If the input is a dataset, return it as is.
- *
- * @param dataset - The dataset to get.
- * @returns The dataset.
- */
-async function getDataset({
-  dataset,
-  client,
-}: {
-  dataset: Dataset | string | Example[];
-  client: PhoenixClient;
-}): Promise<Dataset> {
-  if (typeof dataset === "string") {
-    const datasetResponse = await client
-      .GET(`/v1/datasets/{id}`, { params: { path: { id: dataset } } })
-      .then((d) => d.data?.data);
-    invariant(datasetResponse, `Dataset ${dataset} not found`);
-    const examples = await client
-      .GET(`/v1/datasets/{id}/examples`, { params: { path: { id: dataset } } })
-      .then((e) => e.data?.data);
-    invariant(examples, `Examples for dataset ${dataset} not found`);
-    const datasetWithExamples: Dataset = {
-      ...datasetResponse,
-      examples: examples.examples.map((example) => ({
-        ...example,
-        updatedAt: new Date(example.updated_at),
-      })),
-      versionId: examples.version_id,
-    };
-    return datasetWithExamples;
-  }
-  if (Array.isArray(dataset)) {
-    throw new Error("TODO: implement dataset creation from examples");
-  }
-  return dataset;
 }
 
 /**
