+++ js/packages/phoenix-client/src/index.ts
@@ -1,3 +1,1 @@
 export * from "./client";
-
-export * from "./experiments";
