+++ js/packages/phoenix-client/src/types/logger.ts
@@ -1,0 +1,5 @@
+export type Logger = {
+  info: (message: string) => void;
+  error: (message: string) => void;
+  log: (message: string) => void;
+};
