+++ js/packages/phoenix-client/src/utils/noopLogger.ts
@@ -1,0 +1,7 @@
+import { Logger } from "../types/logger";
+
+export const noopLogger = {
+  info: () => undefined,
+  error: () => undefined,
+  log: () => undefined,
+} satisfies Logger;
